feat(usuario): adiciona cadastro automático de usuário ao cadastrar pessoa

+++ src/main/java/com/api/controleverbasbackend/dto/usuario/DadosCadastroUsuario.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
 public record DadosCadastroUsuario(
-        @NotBlank(message = "CPF é obrigatório") @Pattern(regexp = "\\d{3}\\.?\\d{3}\\.?\\d{3}\\-?\\d{2}", message = "Formato inválido") String cpfPessoa,
-        @NotBlank(message = "Senha é obrigatório") String senha,
-        @NotNull(message = "Não pode ser vazio") Long idTipousuario) {
+                @NotBlank(message = "CPF é obrigatório") @Pattern(regexp = "\\d{3}\\.?\\d{3}\\.?\\d{3}\\-?\\d{2}", message = "Formato inválido") String cpfPessoa,
+                @NotBlank(message = "Senha é obrigatório") String senha,
+                @NotNull(message = "Não pode ser vazio") Integer idTipousuario) {
 }
